Move confluent_kafka to seperate module
