+++ logprep/connector/confluent_kafka/confluent_kafka_input.py
@@ -11,7 +11,7 @@ from socket import getfqdn
 from confluent_kafka import Consumer
 
 from logprep.connector.connector_factory_error import InvalidConfigurationError
-from logprep.connector.confluent_kafka_shared import (
+from logprep.connector.confluent_kafka.confluent_kafka_shared import (
     ConfluentKafka,
     ConfluentKafkaFactory,
     UnknownOptionError,
