+++ logprep/connector/connector_factory.py
@@ -11,11 +11,14 @@ from logprep.output.output import Output
 from logprep.input.dummy_input import DummyInput
 from logprep.input.jsonl_input import JsonlInput
 from logprep.input.json_input import JsonInput
-from logprep.connector.confluent_kafka_input import ConfluentKafkaInput, ConfluentKafkaInputFactory
+from logprep.connector.confluent_kafka.confluent_kafka_input import (
+    ConfluentKafkaInput,
+    ConfluentKafkaInputFactory,
+)
 from logprep.output.dummy_output import DummyOutput
 from logprep.output.writing_output import WritingOutput
 from logprep.output.es_output import ElasticsearchOutput, ElasticsearchOutputFactory
-from logprep.connector.confluent_kafka_output import (
+from logprep.connector.confluent_kafka.confluent_kafka_output import (
     ConfluentKafkaOutput,
     ConfluentKafkaOutputFactory,
 )
