+++ tests/acceptance/util.py
@@ -8,8 +8,12 @@ from multiprocessing import Lock
 from os import path, makedirs
 from os.path import join
 
-from logprep.connector.confluent_kafka_input import ConfluentKafkaInputFactory
-from logprep.connector.confluent_kafka_output import ConfluentKafkaOutputFactory
+from logprep.connector.confluent_kafka.confluent_kafka_input import (
+    ConfluentKafkaInputFactory,
+)
+from logprep.connector.confluent_kafka.confluent_kafka_output import (
+    ConfluentKafkaOutputFactory,
+)
 from logprep.framework.pipeline import Pipeline, SharedCounter
 from logprep.util.helper import recursive_compare, remove_file_if_exists
 from logprep.util.json_handling import parse_jsonl
