+++ tests/unit/connector/test_connector_factory.py
@@ -16,11 +16,15 @@ from logprep.connector.connector_factory import (
     ConnectorFactory,
     UnknownConnectorTypeError,
 )
-from logprep.connector.confluent_kafka_input import ConfluentKafkaInput
+from logprep.connector.confluent_kafka.confluent_kafka_input import (
+    ConfluentKafkaInput,
+)
 from logprep.input.dummy_input import DummyInput
 from logprep.input.json_input import JsonInput
 from logprep.input.jsonl_input import JsonlInput
-from logprep.connector.confluent_kafka_output import ConfluentKafkaOutput
+from logprep.connector.confluent_kafka.confluent_kafka_output import (
+    ConfluentKafkaOutput,
+)
 from logprep.output.dummy_output import DummyOutput
 from logprep.output.es_output import ElasticsearchOutput
 from logprep.output.writing_output import WritingOutput
